WebView: Split into DownloadManager
Create a DownloadManager class which publishes state changes when
something is downloading. This class adds a parent WebViewModel
since error handling is handled there.

The parent is set on the onAppear handler. Thanks to Mantton for helping
out with this commit.
